corrigir os valores de int para double

+++ src/Controle.java
@@ -1,8 +1,8 @@
 public class Controle {
-    static boolean ValidarDeposito(int valor) {
+    static boolean ValidarDeposito(double valor) {
         return valor >= 0;
     }
-    static boolean ValidarTransferencia(Usuario cliente, int valor) {
+    static boolean ValidarTransferencia(Usuario cliente, double valor) {
         return cliente.saldo <= valor;
     }
 }
